+++ private-templates-service/src/config/config.ts
