+++ private-templates-service/src/util/logger.ts
@@ -1,0 +1,8 @@
+import bunyan from 'bunyan';
+
+// Logger
+var logger = bunyan.createLogger({
+    name: 'Adaptive Cards Admin Portal'
+});
+
+export default logger;
